+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-08-21
+### Added
+- Examples in go documentation for [pkg.go.dev/github.com/JamesChung/fofi](https://pkg.go.dev/github.com/JamesChung/fofi)
+- `examples/` directory with runnable examples of each function
+- New `GenerateInputOutputBroadcasters` function
+
+### Changed
+- Renamed `GenerateBroadcasters` to `GenerateOutputBroadcasters`
+
 ## [0.1.0] - 2022-08-20
 ### Added
 - CHANGELOG.md